refactor: move logic from router to service
